File rename only of manifests to maps
git mv does not actually force git to recognize a file move - it will fail if
the file contents also change a great deal in the same commit. Do a preliminary
commit with the bare minimum changes to change the file names.
