+++ mapd/ngnfs-mapd.c
@@ -21,7 +21,7 @@
 #include "shared/lk/err.h"
 #include "shared/lk/kernel.h"
 #include "shared/log.h"
-#include "shared/manifest.h"
+#include "shared/map.h"
 #include "shared/msg.h"
 #include "shared/mtr-socket.h"
 #include "shared/nerr.h"
