+++ shared/map.h
